+++ SmartWeights/SmartWeights/pet-customize-page/Inventory_Customize_page.swift
@@ -7,48 +7,6 @@
 
 import SwiftUI
 
-struct Accessory: Identifiable {
-    var id = UUID()
-    var name: String
-    var imageName: String
-}
-
-struct BackgroundImage: Identifiable {
-    var id = UUID()
-    var name: String
-    var imageName: String
-}
-
-struct Pet_selection: Identifiable {
-    var id = UUID()
-    var name: String
-    var imageName: String
-}
-
-class CustomizeViewModel: ObservableObject {
-    @Published var equippedAccessory: Accessory?
-    @Published var equippedBackgroundImage: BackgroundImage?
-    @Published var equippedPet: Pet_selection? = Pet_selection(name: "Dog", imageName: "Dog") // Default pet
-    @Published var backgroundColor: Color = .white // Default background color
-    
-    // Data arrays
-    let accessories: [Accessory] = [
-        Accessory(name: "Chain", imageName: "chain"),
-        Accessory(name: "Glasses", imageName: "glasses"),
-        Accessory(name: "Jet Pack", imageName: "jetpack"),
-    ]
-    
-    let backgroundImages: [BackgroundImage] = [
-        BackgroundImage(name: "Bamboo", imageName: "Bamboo"),
-        BackgroundImage(name: "Festive", imageName: "Festive"),
-        BackgroundImage(name: "Royal", imageName: "Royal"),
-    ]
-    
-    let pets: [Pet_selection] = [
-        Pet_selection(name: "Dog", imageName: "Dog"),
-        Pet_selection(name: "Cat", imageName: "Cat"),
-    ]
-}
 
 struct Customize_page: View {
     @Environment(\.presentationMode) var presentationMode
